Made significant progress on ButtonInputDetector.
Control can be attached as GestureRecognizer.
Input detection works when control is implemented as view, but not as GestureRecognizer. I'm not sure if what I'm trying to do is actually possible, but I am hopeful, and there are several avenues left to explore:
I'm still hopeful that a compination of PlatformElement and GestureRecognizer is doable. Now that this is working as a view, I know I have everything "wired up" properly, it's just a matter of figuring out how/if that PlatformElement can be used.
If nothing else, input detection can be implemented as an invisible view layer that fakes gesture recognition. It's... definitely not the cleanest solution, but I could definitely do worse.