ButtonInputDetector working as intended on Android with a couple of exceptions:
If attached to a grid at the parent level where the grid's width is not specified, the starting left coordinate of the window is detected as the view's left position. Presumably this can be worked around by figuring out which native controle the Grid layout maps to and then writing code to measure the control's children to get the width. Either way, it works with the ButtonPARCEL as intended for now.

More importantly: still need to write code to cancel down press on exit. That shouldn't be too difficult, though.

After that the exit code for the MaciOS native behavior needs to be written and then we can move on to windows.