Finished ButtonInputDetector implementation for Android and iOS.
May leave issues with grid layout mentioned in previous commit in place, as they appear to be universal and they make a certain sense: if the grid is centered and has no specified width, why wouldn't it expand to the edges. And if the behavior is attached to the grid, why wouldn't it account for that empty space?
Need to build implementation for Windows and also resolve the rendering inconsistencies between platforms via a custom handler implementation called during the constructor.