Implements strategy pattern for GaugePARCEL.
The GaugePARCEL class was becoming way to monolithic. There were dense conditionals in the renderer and parent that evaluated the same values and were adding a lot of complexity. To address this, the renderers and behaviors have been moved into strategy objects. The GaugePARCEL renderer can still be set separately, so this doesn't reduce control extendability. The renderer in each strategy is initialized at the time it's first requested, which never occurs if the ControlPARCEL's Renderer property is defined, so there is no additional overhead involved beyond the additional classes.

Additionally, this actually provided the ability to improve control customizability: The abstract strategy base class is publicly available for extending and GaugePARCEL's strategy is implemented as a bindable property. Previously GaugePARCEL's behavior was coupled to the value of the Appearance property. Now, because the various strategies also encapsulate the control's behavior, new strategies can be created and assigned to the control and the control can still be used with either a custom renderer or one of the included ones. Simply put: This makes behavior itself customizable for GaugePARCEL.