Cleaned up some of the existing strategy behavior and began laying the groundwork for a few things.
The ultimate goal with GaugePARCEL at this point should be to get rid of the existing enum and move all of that functionailty entirely into the new strategies. This creates a complication, I overlooked before, however: Currently, the strategies do not have a parameterless constructor... I suspect that, at the very least, this can be worked around via a valueCoercion method, but that seems like an inelegant solution.